backend/pyproject.toml
[project]
name = "flask_sample"
version = "0.1.0"
description = "Sample Flask application"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi[standard]>=0.115.12",
    "flask==3.1.0",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.4.2",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.24",
    "matplotlib>=3.10.3",
    "notebook>=7.4.2",
    "numpy>=2.2.6",
    "openai>=1.86.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.12.0",
    "ty>=0.0.1a11",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
